refactor: lib.rs divided into contract related modules
