+++ src/modules/mod.rs
@@ -1,0 +1,9 @@
+mod init_maps;
+
+mod graph_token;
+
+mod staking;
+
+mod curation;
+
+mod graph_out;
